refactor : 디렉토리 구조 변경
util 함수들을 utils 디렉토리를 생성 후 utils 디렉토리로 이동.
